Split alot of MameServer functionality into 'Broker' since I'm about to reuse a bunch of it
